test: make both tox and travis use the same test-requirements.txt file
This should fix build errors that were being seen by Steve on
this branch.  Since tox creates a unique environment for each
interpreter, you need to give some hints as to what to needs to be
installed in each environment.  There are a few items which are overkill
(tox itself, nose) that probably wouldn't need to be done in the tox
environment.  At the same time, I wouldn't expect those extra deps
from breaking anything.