docs: minor improvements to the documentation
The only substantive change here is a clarification that the framework
technically provides dependency injection not inversion of control.  The
two are not unrelated, but it is more correct to use dependency injection
here.

This change also fixes the link to the build status on travis when
the repo was moved from the etherios organization to digidotcom.

Signed-off-by: Paul Osborne <Paul.Osborne@digi.com>